+++ ESFrameworkUI/ESBadgeView.m
@@ -280,7 +280,7 @@
 - (void)drawRect:(CGRect)rect
 {
     CGContextRef context = UIGraphicsGetCurrentContext();
-    if (self.isFlatStyle || (self.applyFlatStyleAutomatically && ESOSVersionIsAtLeast(7))) {
+    if (self.isFlatStyle || (self.applyFlatStyleAutomatically && ESOSVersionIsAtLeast(7, 0))) {
         [self drawFlatStyleWithContext:context withRect:rect];
     } else {
         self.cornerRoundness = 0.4;